Fixed naming problem and added test case to database tests.
