+++ UNO/Database/CreateAccountData.java
@@ -2,4 +2,27 @@ package Database;
 
 public class CreateAccountData {
 
+    private String username;
+    private String password;
+
+    public String getUsername()
+    {
+        return username;
+    }
+
+    public String getPassword()
+    {
+        return password;
+    }
+
+    public void setUsername(String username)
+    {
+        this.username = username;
+    }
+
+    public void setPassword(String password)
+    {
+        this.password = password;
+    }
+
 }
