--- UNO/Database/CreateUserAccount.java
@@ -1,28 +1,0 @@
-package Database;
-
-public class CreateUserAccount {
-
-    private String username;
-    private String password;
-
-    public String getUsername()
-    {
-        return username;
-    }
-
-    public String getPassword()
-    {
-        return password;
-    }
-
-    public void setUsername(String username)
-    {
-        this.username = username;
-    }
-
-    public void setPassword(String password)
-    {
-        this.password = password;
-    }
-
-}
